# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis/{symbol}:
    post:
      operationId: analyzeStock
      tags: [analysis]
      summary: Generate AI analysis for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI-generated analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockAnalysis"
        "404":
          description: Stock not found
...
  /analysis/{symbol}/longterm:
    post:
      operationId: analyzeLongTerm
      tags: [analysis]
      summary: AI long-term investment thesis for a stock
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Long-term investment analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LongTermAnalysis"
        "404":
          description: Stock not found
  /analysis/{symbol}/swing:
    post:
      operationId: analyzeSwing
      tags: [analysis]
      summary: AI swing trade setup and technical analysis
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Swing trade technical analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwingAnalysis"
        "404":
          description: Stock not found
...
components:
  schemas:
...
    LongTermAnalysis:
      type: object
      properties:
        symbol:
          type: string
        thesis:
          type: string
        growthCatalysts:
          type: array
          items:
            type: string
        longTermRisks:
          type: array
          items:
            type: string
        valuationView:
          type: string
        futureOutlook:
          type: string
        sentiment:
          type: string
          enum: ["bullish", "bearish", "neutral"]
        priceTarget1Y:
          type: number
          nullable: true
        priceTarget3Y:
          type: number
          nullable: true
        generatedAt:
          type: string
      required:
        - symbol
        - thesis
        - growthCatalysts
        - longTermRisks
        - valuationView
        - futureOutlook
        - sentiment
        - generatedAt
    SwingAnalysis:
      type: object
      properties:
        symbol:
          type: string
        setup:
          type: string
        trend:
          type: string
          enum: ["uptrend", "downtrend", "sideways"]
        momentum:
          type: string
          enum: ["strong_bullish", "bullish", "neutral", "bearish", "strong_bearish"]
        keySupport:
          type: number
          nullable: true
        keyResistance:
          type: number
          nullable: true
        entryZone:
          type: string
        targetPrice:
          type: number
          nullable: true
        stopLoss:
          type: number
          nullable: true
        timeframe:
          type: string
        riskReward:
          type: string
        keyWatchLevels:
          type: array
          items:
            type: string
        tradingNotes:
          type: string
        generatedAt:
          type: string
      required:
        - symbol
        - setup
        - trend
        - momentum
        - entryZone
        - timeframe
        - riskReward
        - keyWatchLevels
        - tradingNotes
        - generatedAt
    StockAnalysis:
      type: object
      properties:
        symbol:
          type: string
        summary:
          type: string
        sentiment:
          type: string
          enum: ["bullish", "bearish", "neutral"]
        sentimentScore:
          type: number
        keyStrengths:
          type: array
          items:
            type: string
        keyRisks:
          type: array
          items:
            type: string
        technicalOutlook:
          type: string
        fundamentalOutlook:
          type: string
        recommendation:
          type: string
          enum: ["strong_buy", "buy", "hold", "sell", "strong_sell"]
        priceTarget:
          type: number
          nullable: true
        generatedAt:
          type: string
      required:
        - symbol
        - summary
        - sentiment
        - sentimentScore
        - keyStrengths
        - keyRisks
        - technicalOutlook
        - fundamentalOutlook
        - recommendation
        - generatedAt
...